+++ src/hooks/useMovies.ts
@@ -1,0 +1,75 @@
+import { useState, useEffect, useMemo } from 'react';
+import { type Movie } from '../types/Movie';
+
+export const useMovies = () => {
+  // 상태들
+  const [movies, setMovies] = useState<Movie[]>([]);
+  const [isLoading, setIsLoading] = useState<boolean>(false);
+  const [error, setError] = useState<string | null>(null);
+  const [searchTerm, setSearchTerm] = useState<string>("");
+
+  // 영화 조회하기
+  useEffect(() => {
+    const fetchMovies = async () => {
+      setIsLoading(true);
+      try {
+        const res = await fetch("/api/movies");
+        if (!res.ok) throw new Error("영화 데이터를 불러오지 못했습니다.");
+        const data: Movie[] = await res.json();
+        setMovies(data);
+      } catch (err: unknown) {
+        if (err instanceof Error) {
+          setError(err.message);
+        } else {
+          setError(String(err));
+        }
+      } finally {
+        setIsLoading(false);
+      }
+    };
+    fetchMovies();
+  }, []); // 컴포넌트가 처음 마운트될 때 한 번만 실행
+
+  // 영화 검색하기
+  const filteredMovies = useMemo(() => {
+    return movies.filter((movie) =>
+      movie.title.toLowerCase().includes(searchTerm.toLowerCase())
+    );
+  }, [movies, searchTerm]); // searchTerm이 바뀔 때마다 자동으로 실행
+
+  // 영화 추가하기
+  const addMovie = async (newMovie: Omit<Movie, 'id'>) => {
+    try {
+      setIsLoading(true);
+      const res = await fetch("/api/movies", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify(newMovie),
+      });
+
+      if (!res.ok) throw new Error("영화를 추가하지 못했습니다.");
+
+      const savedMovie: Movie = await res.json();
+      setMovies((prev) => [...prev, savedMovie]);
+      setError(null);
+    } catch (err: unknown) {
+      if (err instanceof Error) {
+        setError(err.message);
+      } else {
+        setError(String(err));
+      }
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  return {
+    movies: filteredMovies,
+    isLoading,
+    error,
+    searchTerm,
+    setSearchTerm,
+    addMovie,
+    setError
+  };
+};
